requirements.txt
# Without submodules
schedule==1.1.0
requests==2.27.1
pyqt5==5.15.4
flask==2.0.2
pyinstaller==4.7
flask-wtf==1.0.0
pygame==2.1.0
sqlalchemy==1.4.31
Pillow==8.3.2
flask-login==0.5.0
sqlalchemy-serializer==1.4.1
pytest==7.0.1
alembic==1.7.6
flask-restful==0.3.9
python-telegram-bot==13.11
discord.py==1.7.3
pymorphy2==0.9.1
gevent==21.12.0
git+https://github.com/adamakalshchikov/vk_api.git